improvement para el algoritmo de fibonacci

+++ src/FibonacciWebApi/Services/IFibonacciService.cs
@@ -7,7 +7,7 @@ namespace FibonacciWebApi.Services
 {
      public interface IFibonacciService
     {
-         int CalcularFibo(int num);
+         long CalcularFibo(int num);
 
     }
 }
